Run mapping on translated rerun at the time of creation
