+++ openedx/features/wikimedia_features/meta_translations/urls.py
@@ -3,14 +3,14 @@ Urls for Meta Translations
 """
 from django.conf.urls import include, url
 
-from  openedx.features.wikimedia_features.meta_translations.views import course_blocks_mapping, render_translation_home, update_block_direction_flag
+from  openedx.features.wikimedia_features.meta_translations.views import course_blocks_mapping_view, render_translation_home, update_block_direction_flag
 
 app_name = 'meta_translations'
 
 urlpatterns = [
     url(
         r'^course_blocks_mapping/$',
-        course_blocks_mapping,
+        course_blocks_mapping_view,
         name='course_blocks_mapping'
     ),
     url(
